requirements.txt
discord_markdown_ast_parser
discord.py
python-dotenv